server/requirements.txt
django-mongodb-backend
django-cors-headers
# react-router-dom